[3/N][Attention] Move AttentionMetadata-related code from utils.py to backend.py (#32054)
